Added python script to normalize a sketch.
Running this script on a sketch (or series of sketches) will scale,
rotate, and translate the sketch so that the resetting the
transform will bring you back to the initial size, orientation, and
position. In other words, the initial scene transformation becomes
identity. The environment will change positions as a result.

Change-Id: Id0b27446a824f8712f05498c5a317020a32f1645