glb2: Implement compatibility mode
This is the other half of go/tbcl/109834, which renames attributes
from (for example) TEXCOORD_n -> _TB_UNITY_TEXCOORD_n.

At import time, rename them back to TEXCOORD_n because that what the
rest of the importer expects. If there happens to already be a
TEXCOORD_n, that accessor is ignored (which is the desired behavior;
we don't want or need to bring in the "compatible" texcoord data)

Change-Id: Id000b015a0ee4b1f702b2f1e6354934ba376cd46